requirements.txt
celery==5.3.4
redis==4.6.0
pymongo==4.5.0
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.7 